Created web page to allow user to upload SQLite db file to be imported.
The HTTP endpoint that that the web form uploads to does some validation
for the existence of a file upload, that the file is of the correct
type, and then it reads the SQLite db file into the system.

Upon reading the file into the system it validates the data within the
two database tables before copying that data into the cloud database.