Created web page to allow user to send SQL queries to the cloud db.
The page contains a web form that accepts an SQL query (only SELECT
statements) and it sends the request to an HTTP endpoint which does some
validation on the query itself.

If the presence of a SELECT SQL query is successfully validated then it is
made to the cloud database. If for any reason the SELECT SQL query is
badly formed then an error will be returned to the web form page.

If the query is successfully made, whether it contains non-empty results
or not it will be displayed on the page containing the web form along
with a message stating whether the results were empty or there were was
data returned.

If there is data returned then the results will be displayed in a table
which is powered by VueJS (a JavaScript framework) for a nicer user
experience that offers filtering, sorting and pagination capabilities.